Fix deadlock in stopAllThreads() (#61)
stopAllThreads() and createThread() can deadlock (and sometimes did)
because the first function takes out the threadMapMux lock and then
tries to pthread_join() other threads.

createThread() also tries to take out the lock but if stopAllThreads()
owns it and another thread calls createThread(), then boom, the threads
would deadlock each other.  Drop the lock before calling pthread_join()
and reacquire it afterwards.

This problem was noticed on an AIX machine with Node.js Application
Metrics but since the Linux port uses identical logic, I've updated
that as well.  The MacOS port has that code commented out but that
didn't stop me from updating it anyway (but leaving it commented out.)