update omr-agentcore to not assume thread is running in char mode (#95)
* Fix native2Ascii conversion when building without -qascii

native2Ascii() needs to result in an ASCII string regardless of -qascii compiler
flag. Currently there's an issue in that native2Ascii() no longer does any
conversion when building without -qascii. This is because expectedNativeCodepage()
returns "IBM-1047", and thus the call to convertCodePage() ends up returning 0
which prevented the __etoa() subsequent call.

The fix here is to revert the changes in 182b635 and instead convert to ASCII only
if the thread's character mode is EBCDIC.

* move call to __etoa outside the if convertToCurrentLocale arg

The arg is supposed to convert the incoming str from user defined locale to IBM-1047, however the native2Ascii API needs to result in an ASCII str in the end, irregardless of what the native locale is.

* remove call to node's dprintf()

It's not needed since -qascii was added to omr-agentcore and appmetrics